fix cuda compile bug
