+++ source/module_hamilt_lcao/module_gint/gtask_rho.cpp
@@ -1,6 +1,6 @@
 #include "gint_rho_gpu.h"
 #include "module_base/ylm.h"
-#include "module_hamilt_lcao/module_gint/gint_tools.h"
+#include "gint_tools/gint_tools.h"
 #include "omp.h"
 namespace GintKernel
 {
